обновит компоненты
