+++ src/pages/main-page/main-page.tsx
@@ -1,13 +1,30 @@
-import { Helmet } from 'react-helmet-async';
+import { useState } from 'react';
+import { useCallback } from 'react';
 
-import Header from '../../components/header/header';
-import Cities from '../../components/cities/cities';
-import Location from '../../components/location/location';
-import { useAppSelector } from '../../hooks';
-import { getActiveCity, getOffers, getSortItem, sortOffers } from '../../store/selectors';
+import { Helmet } from 'react-helmet-async';
 import classNames from 'classnames';
 
+import Header from '../../components/header/header';
+import Location from '../../components/location/location';
+import OfferListMemo from '../../components/offer-list/offer-list';
+import SortListMemo from '../../components/sort-list/sort-list';
+import MapMemo from '../../components/map/map';
+import NoCards from '../../components/no-cards/no-cards';
+
+import { TOffer } from '../../types/offer';
+import { useAppSelector, useAppDispatch } from '../../hooks';
+import { addPluralEnding } from '../../utils/utils';
+import { TSortItem } from '../../types/sort';
+
+import { getActiveCity, getSortItem, sortOffers } from '../../store/app-process/selectors';
+import { setActiveSortItem } from '../../store/app-process/app-process';
+import { getOffers } from '../../store/data-process/selectors';
+
 function MainPage(): JSX.Element {
+
+  const dispatch = useAppDispatch();
+
+  const [selectedPointId, setSelectedPointId] = useState<TOffer['id'] | null>(null);
 
   const activeSortItem = useAppSelector(getSortItem);
 
@@ -17,7 +34,16 @@ function MainPage(): JSX.Element {
 
   const filteredOffers = offers.filter((offer) => offer.city.name === activeCity.name);
 
-  const currentOffers = useAppSelector(() => sortOffers({ offers: filteredOffers, activeSortItem }));
+  const currentOffers = useAppSelector((state) => sortOffers(state, filteredOffers));
+
+  function handleListItemHover(itemId: TOffer['id'] | null) {
+    setSelectedPointId(itemId);
+  }
+
+  const handleSortItems = useCallback((type: TSortItem) => {
+    dispatch(setActiveSortItem(type));
+    return currentOffers;
+  }, [currentOffers, dispatch]);
 
   return (
     <div className="page page--gray page--main">
@@ -31,10 +57,38 @@ function MainPage(): JSX.Element {
         { 'page__main--index-empty': currentOffers.length === 0 })}
       >
         <h1 className="visually-hidden">Cities</h1>
-        <div className="tabs">
-          <Location />
+        <Location />
+        {/* <Cities /> */}
+        <div className="cities">
+          {currentOffers.length === 0 ? <NoCards /> :
+            <div className="cities__places-container container">
+              <section className="cities__places places">
+                <h2 className="visually-hidden">Places</h2>
+                <b className="places__found">{currentOffers.length} place{addPluralEnding(currentOffers.length)} to stay in {activeCity.name}</b>
+                <SortListMemo
+                  activeSortItem={activeSortItem}
+                  onSortItems={handleSortItems}
+                />
+
+                <OfferListMemo
+                  offers={currentOffers}
+                  block={'cities'}
+                  onListItemHover={handleListItemHover}
+                />
+
+              </section>
+              <div className="cities__right-section">
+
+                <MapMemo
+                  block={'cities'}
+                  offers={offers}
+                  location={activeCity.location}
+                  selectedPointId={selectedPointId}
+                />
+
+              </div>
+            </div>}
         </div>
-        <Cities />
       </main>
     </div>
   );
