+++ src/types/offer-preview.ts
@@ -3,6 +3,7 @@ import { TLocation } from './location';
 
 export type TOfferPreview = {
   city: TCity;
+  id: string;
   isFavorite: boolean;
   isPremium: boolean;
   location: TLocation;
@@ -11,9 +12,6 @@ export type TOfferPreview = {
   rating: number;
   title: string;
   type: string;
-  id: string;
-  // onCardHover?: void;
-  // block?: string;
  }
 
 export type TOfferPreviews = TOfferPreview[];
